Sagheer Baloch
This paper will present a critical review of the Baloch and Balochistan showing that how the Baloch has evolved to become a nation, and Balochistan from being a hinterland in Iranian plateau to a province of Pakistani federation. The paper will discuss the origin of the Baloch and examine the various theories about the origin of the Baloch nation and argue that neither the primordialist nor the modernist theories of nationhood can explain the Baloch nation. In fact, it is the ethno-symbolists that offer the best understanding of the Baloch as a nation. The Baloch nation is heterogeneous in its composition, being an admixture of various ethnic, racial, and linguistic groups over a long period of time. Baloch nationalism, however, is best understandable via the modernist theories of nationalism. The paper will contend that the Baloch nation is neither ancient nor modern because it emerged as a nation before the era of enlightenment and the French revolution. The paper will elaborate the modern Baloch nation, the birth of ‗Baloch Confederacy' in its historical perspective, from the arrival of the British in 1839 to the fall of the State of Kalat into the Pakistani federation in 1948. The paper will finally provide a brief but robust analysis on the emergence, evolution and dynamics of contemporary Baloch nationalism vis-à-vis the federation of Pakistan